Changed connection class variable names in SQL.cs
Sorry if I broke anything but I went in and changed the variable names from camelcase to lowercase with dashes. Reason being that Matt, Loren and I have agreed on having that be the standard throughout the code and I would like to keep everything similar for ease of reading.